Fix .off() for multiple elements
I'm having some trouble with `$el.off()` - that is to say it is not detaching event listeners from multiple elements.

When `$elements.on(type, handler)` is called, each element in `$elements` gets assigned a `__domtastic_event__` attributes, which maps it to an event handler domtastic's internal `handlers` array.

However, when `$.off(type, handler)` is called with the same type and handler, this `__domtastic_event__` is not unset - in fact its value is used as a key on the element, against which `null` is set. That is to say, if `element.__domtastic__event__` was set to `2`, then after `clearHandlers` was run, `element.__domtastic_event__` would still be `2`, and `element['2']` is `null`.

I have no idea how your [existing test](https://github.com/webpro/DOMtastic/blob/master/test/spec/event.js#L344-L351) for this is passing, as when I run it locally it fails.